refactor(runs): split controller responsibilities
