+++ pkg/runs/attach_output.go
@@ -3,6 +3,7 @@ package runs
 import (
 	"time"
 
+	driverpkg "agent-compose/pkg/driver"
 	domain "agent-compose/pkg/model"
 )
 
@@ -39,3 +40,38 @@ type RunAttachOutput struct {
 }
 
 type RunAttachSender func(RunAttachOutput) error
+
+func runAttachStartedResponse(run domain.ProjectRunRecord, sandbox *domain.Sandbox, warnings []string, startedAt time.Time) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputStarted, CreatedAt: startedAt, Run: run, SandboxID: sandbox.Summary.ID, Warnings: append([]string(nil), warnings...)}
+}
+
+func runAttachOutputResponse(data []byte, stream domain.StdioStream, tty bool) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputData, CreatedAt: time.Now().UTC(), Data: append([]byte(nil), data...), Stream: stream, TTY: tty}
+}
+
+func runAttachAgentEventResponse(name, text, payloadJSON string) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputAgentEvent, CreatedAt: time.Now().UTC(), Name: name, Text: text, PayloadJSON: payloadJSON}
+}
+
+func runAttachAgentTurnCompletedResponse(run domain.ProjectRunRecord, resultJSON string, warnings []string) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputAgentTurnCompleted, CreatedAt: time.Now().UTC(), Run: run, ResultJSON: resultJSON, Warnings: append([]string(nil), warnings...)}
+}
+
+func runAttachResultResponse(run domain.ProjectRunRecord, transition TransitionRequest, success bool) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputResult, CreatedAt: time.Now().UTC(), Run: run, ExitCode: transition.ExitCode, Success: success, Output: transition.Output, ResultJSON: transition.ResultJSON, Error: transition.Error}
+}
+
+func runAttachErrorResponse(code, message string, terminal bool) RunAttachOutput {
+	return RunAttachOutput{Kind: RunAttachOutputError, CreatedAt: time.Now().UTC(), Code: code, Error: message, Terminal: terminal}
+}
+
+func driverOutputStreamToRun(frameType driverpkg.RuntimeOutputFrameType) domain.StdioStream {
+	if frameType == driverpkg.RuntimeOutputStderr {
+		return domain.StdioStderr
+	}
+	return domain.StdioStdout
+}
+
+func warningsFromRun(run domain.ProjectRunRecord) []string {
+	return append([]string(nil), run.Warnings...)
+}
