+++ pkg/runs/command_execution.go
@@ -1,0 +1,143 @@
+package runs
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"time"
+
+	appconfig "agent-compose/pkg/config"
+	"agent-compose/pkg/execution"
+	domain "agent-compose/pkg/model"
+	agentcomposev2 "agent-compose/proto/agentcompose/v2"
+)
+
+func (c *Controller) executeProjectRunCommand(ctx context.Context, run domain.ProjectRunRecord, sandbox *domain.Sandbox, req RunAgentRequest, commandText string, sink *StreamSink) (TransitionRequest, error) {
+	artifactsDir := projectRunCommandArtifactsDir(run, sandbox)
+	logsPath := filepath.Join(artifactsDir, "transcript.txt")
+	transition := TransitionRequest{
+		RunID:     run.RunID,
+		SandboxID: sandbox.Summary.ID,
+		LogsPath:  logsPath,
+	}
+	if c.store == nil || c.runtime == nil {
+		err := fmt.Errorf("command runtime dependencies are required")
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	if sink != nil && sink.SendStarted != nil {
+		if err := sink.SendStarted(run, time.Now().UTC()); err != nil {
+			transition.ExitCode = 1
+			transition.Error = fmt.Sprintf("command execution failed: %v", err)
+			return transition, err
+		}
+	}
+	appconfig.ApplyDefaultGuestPaths(c.config)
+	vmState, err := c.store.GetVMState(sandbox.Summary.ID)
+	if err != nil {
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	runtime, err := c.runtime(sandbox)
+	if err != nil {
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	if err := os.MkdirAll(artifactsDir, 0o755); err != nil {
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	guestArtifactsDir := filepath.Join(c.config.GuestStateRoot, "runs", run.RunID)
+	runtimeRequest := execution.RuntimeCommandRequestPayloadFromCommand(
+		c.config,
+		"shell",
+		"",
+		nil,
+		commandText,
+		c.config.GuestWorkspacePath,
+		execEnvMap(req.Env),
+		0,
+		0,
+		guestArtifactsDir,
+	)
+	if err := execution.WriteJSONArtifact(filepath.Join(artifactsDir, "command-request.json"), runtimeRequest); err != nil {
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	var sendErr error
+	writer := func(chunk domain.ExecChunk) {
+		if sendErr != nil {
+			return
+		}
+		filtered, visible := execution.FilterCommandStreamChunk(chunk)
+		if !visible {
+			return
+		}
+		offset, err := appendProjectRunLogChunk(logsPath, filtered)
+		if err != nil {
+			sendErr = err
+			return
+		}
+		c.publishRunLogChunk(run.RunID, filtered, offset)
+		if sink != nil && sink.SendChunk != nil {
+			sendErr = sink.SendChunk(run.RunID, filtered, time.Now().UTC())
+		}
+	}
+	execCtx, cancel := execution.ExecContext(ctx, 0)
+	defer cancel()
+	result, execErr := runtime.ExecStream(execCtx, sandbox, vmState, execution.BuildRuntimeCommandExecSpec(c.config, sandbox, filepath.Join(guestArtifactsDir, "command-request.json"), c.config.GuestHomePath), writer)
+	if sendErr != nil {
+		transition.ExitCode = 1
+		transition.Error = fmt.Sprintf("command execution failed: %v", sendErr)
+		return transition, sendErr
+	}
+	if execErr != nil {
+		result.ExitCode = execution.FirstNonZeroInt(result.ExitCode, 1)
+		result.Success = false
+		if strings.TrimSpace(result.Output) == "" {
+			result.Output = firstNonEmpty(result.Stderr, result.Stdout, execErr.Error())
+		}
+		transition = transitionFromCommandResult(run, sandbox, commandText, result, execErr)
+		transition.LogsPath = logsPath
+		return transition, execErr
+	}
+	commandResult, err := execution.ParseCommandExecResult(result)
+	if err != nil {
+		transition.ExitCode = execution.FirstNonZeroInt(transition.ExitCode, 1)
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	if err := execution.MirrorRuntimeCommandArtifacts(artifactsDir, commandResult); err != nil {
+		transition.ExitCode = execution.FirstNonZeroInt(commandResult.ExitCode, 1)
+		transition.Error = fmt.Sprintf("command execution failed: %v", err)
+		return transition, err
+	}
+	transition = transitionFromCommandResult(run, sandbox, commandText, execution.RuntimeCommandResultToExecResult(commandResult), nil)
+	transition.LogsPath = logsPath
+	return transition, nil
+}
+
+func execEnvMap(items []*agentcomposev2.EnvVarSpec) map[string]string {
+	if len(items) == 0 {
+		return nil
+	}
+	env := make(map[string]string)
+	for _, item := range items {
+		name := strings.TrimSpace(item.GetName())
+		if name == "" {
+			continue
+		}
+		env[name] = item.GetValue()
+	}
+	if len(env) == 0 {
+		return nil
+	}
+	return env
+}
